requirements.txt
playwright
pandas
Pillow
pywin32